Refactor commit fetching to CommitHistory component
