+++ src/components/DetailsCard.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useId, useMemo, useRef, useState } from "react"
-import { Form, useFetcher, useLocation, useNavigation } from "@remix-run/react"
-import type { HydratedGitBlobObject, HydratedGitObject, HydratedGitTreeObject, GitLogEntry } from "~/analyzer/model"
+import { Form, useLocation, useNavigation } from "@remix-run/react"
+import type { HydratedGitBlobObject, HydratedGitObject, HydratedGitTreeObject } from "~/analyzer/model"
 import { AuthorDistFragment } from "~/components/AuthorDistFragment"
 import { ChevronButton } from "~/components/ChevronButton"
 import { CloseButton } from "~/components/util"
@@ -13,12 +13,11 @@ import byteSize from "byte-size"
 import type { AuthorshipType } from "~/metrics/metrics"
 import { mdiAccountMultiple, mdiOpenInNew, mdiEyeOffOutline, mdiFile, mdiFolder } from "@mdi/js"
 import { Icon } from "@mdi/react"
-import { FileHistoryElement } from "./FileHistoryElement"
+import { CommitHistory } from "./CommitHistory"
 import clsx from "clsx"
 import { useMetrics } from "~/contexts/MetricContext"
 import { MenuItem, MenuTab } from "./MenuTab"
 import { CommitsCard } from "./CommitsCard"
-import type { CommitsPayload } from "~/routes/commits"
 
 function OneFolderOut(path: string) {
   const index = path.lastIndexOf("/")
@@ -42,12 +41,6 @@ export function DetailsCard({
   const { setPath, path } = usePath()
   const { analyzerData } = useData()
   const isProcessingHideRef = useRef(false)
-  const [commitsPayload, setCommitsPayload] = useState<CommitsPayload | null>(null)
-  const [lastClicked, setLastClicked] = useState("")
-  const [commitIndex, setCommitIndex] = useState(0)
-  const commitIncrement = 5
-
-  const fetcher = useFetcher()
 
   useEffect(() => {
     if (isProcessingHideRef.current) {
@@ -55,27 +48,6 @@ export function DetailsCard({
       isProcessingHideRef.current = false
     }
   }, [clickedObject, setClickedObject, state])
-
-  useEffect(() => {
-    if (clickedObject && clickedObject.path !== lastClicked) {
-      setLastClicked(clickedObject.path)
-      setCommitIndex(commitIndex + commitIncrement)
-      fetcher.load(
-        `/commits?path=+${clickedObject.path}&isFile=${
-          clickedObject.type === "blob"
-        }&skip=${commitIndex}&count=${commitIncrement}`,
-      )
-    }
-  }, [clickedObject])
-
-  useEffect(() => {
-    if (fetcher.state === "idle") {
-      setCommitsPayload(fetcher.data)
-    } else if (fetcher.state === "loading") {
-      setCommitIndex(0)
-      setCommitsPayload(null)
-    }
-  }, [fetcher.state])
 
   useEffect(() => {
     // Update clickedObject if data changes
@@ -155,7 +127,7 @@ export function DetailsCard({
           Group authors
         </button>
         <div className="card bg-white/70 text-black">
-          <FileHistoryElement commitsPayload={commitsPayload} commitCutoff={commitIncrement}/>
+          <CommitHistory />
         </div>
       </div>
       <div className="flex gap-2">
@@ -254,9 +226,6 @@ export function DetailsCard({
               <Icon path={mdiAccountMultiple} />
               Group authors
             </button>
-            <div className="card bg-white/70 text-black">
-              <FileHistoryElement commitsPayload={commitsPayload} state={state} clickedObject={clickedObject} />
-            </div>
           </div>
           <div className="mt-2 flex gap-2">
             {isBlob ? (
@@ -323,7 +292,9 @@ export function DetailsCard({
             )}
           </div>
         </MenuItem>
-        <MenuItem title="Commits">{CommitsCard()}</MenuItem>
+        <MenuItem title="Commits">
+          <CommitsCard />
+        </MenuItem>
       </MenuTab>
     </div>
   )
